keyboard shortcuts link always visible
