+++ src/app/components/AccessibilityButtonsWrapper.spec.tsx
@@ -8,7 +8,6 @@ import * as analytics from '../../helpers/analytics';
 import { book } from '../../test/mocks/archiveLoader';
 import { expectError, expectReactRenderError, renderToDom } from '../../test/reactutils';
 import TestContainer from '../../test/TestContainer';
-import { openKeyboardShortcutsMenu } from '../content/keyboardShortcuts/actions';
 import AccessibilityButtonsWrapper from './AccessibilityButtonsWrapper';
 import { HiddenButton } from './HiddenLink';
 import MainContent from './MainContent';
@@ -72,34 +71,5 @@ describe('AccessibilityButtonsWrapper', () => {
 
             expect(spyFocus).toHaveBeenCalledTimes(1);
         }
-    });
-
-    it('opens the Keyboard Shortcuts Menu when that link is clicked', () => {
-        const dispatch = jest.fn();
-        const spyRedux = jest.spyOn(redux, 'useDispatch').mockImplementation(() => dispatch);
-
-        const trackOpenCloseKS = jest.fn();
-        const spyAnalytics = jest.spyOn(analytics, 'useAnalyticsEvent').mockImplementation(
-          () => trackOpenCloseKS
-        );
-
-        const {root} = renderer.create(<TestContainer>
-          <AccessibilityButtonsWrapper/>
-        </TestContainer>);
-
-        expect(spyRedux).toHaveBeenCalledTimes(1);
-
-        expect(spyAnalytics).toHaveBeenCalledTimes(1);
-        expect(spyAnalytics).toHaveBeenCalledWith('openCloseKeyboardShortcuts');
-
-        const preventDefault = jest.fn();
-        root.findByType(HiddenButton).props.onClick({preventDefault});
-
-        expect(preventDefault).toHaveBeenCalledTimes(1);
-
-        expect(dispatch).toHaveBeenCalledTimes(1);
-        expect(dispatch).toHaveBeenCalledWith(openKeyboardShortcutsMenu());
-
-        expect(trackOpenCloseKS).toHaveBeenCalledTimes(1);
     });
 });
