+++ src/app/messages/en/messages.json
@@ -4,7 +4,6 @@
   "i18n:a11y:accessibilityHelp": "Go to accessibility page",
   "i18n:a11y:accessibilityHelp:link": "https://openstax.org/accessibility-statement",
   "i18n:assigned:button:continue": "Mark as done & continue",
-  "i18n:a11y:keyboard-shortcuts:menu": "Keyboard shortcuts menu",
   "i18n:a11y:keyboard-shortcuts:heading": "Keyboard Shortcuts",
   "i18n:a11y:keyboard-shortcuts:close": "Click to close Keyboard Shortcuts menu",
   "i18n:a11y:keyboard-shortcuts:sub-heading": "Using a Keyboard for Navigation",
@@ -17,7 +16,7 @@
   "i18n:a11y:keyboard-shortcuts:deselect-highlight-color": "Deselect current highlight color",
   "i18n:a11y:keyboard-shortcuts:save-or-cancel": "Save or Cancel note edits",
   "i18n:a11y:keyboard-shortcuts:creating-highlights-and-notes": "Creating Highlights and Notes with a Keyboard",
-  "i18n:a11y:keyboard-shortcuts:caret-extension": "This functionality requires the use of a <strong>Caret Browsing extension</strong>, which enables you to browse and select the text of web pages using arrow keys.",
+  "i18n:a11y:keyboard-shortcuts:caret-browsing": "To select text for highlighting or note-taking, enable caret browsing by pressing F7 (or Fn+F7) in your browser. This allows you to browse and select text using the arrow keys, without needing any extension.",
   "i18n:a11y:keyboard-shortcuts:move-focus-through-each-line": "Move focus through each line of text",
   "i18n:a11y:keyboard-shortcuts:select-text": "Select text",
   "i18n:a11y:keyboard-shortcuts:create-highlight-or-note": "Create a highlight or note after selecting text",
