+++ src/app/messages/es/messages.json
@@ -2,7 +2,6 @@
   "i18n:a11y:skipToContent": "Omitir e ir al contenido",
   "i18n:a11y:accessibilityHelp": "Ir a la página de accesibilidad",
   "i18n:a11y:accessibilityHelp:link": "n/a (tbd for now)",
-  "i18n:a11y:keyboard-shortcuts:menu": "Menú de atajos de teclado",
   "i18n:a11y:keyboard-shortcuts:heading": "Atajos de teclado",
   "i18n:a11y:keyboard-shortcuts:close": "Haga clic para cerrar el menú de atajos de teclado",
   "i18n:a11y:keyboard-shortcuts:sub-heading": "Cómo usar el teclado para navegar",
@@ -15,7 +14,7 @@
   "i18n:a11y:keyboard-shortcuts:deselect-highlight-color": "Desmarcar el color de resaltado actual",
   "i18n:a11y:keyboard-shortcuts:save-or-cancel": "Guardar o cancelar cambios a una nota",
   "i18n:a11y:keyboard-shortcuts:creating-highlights-and-notes": "Cómo resaltar y agregar notas con el teclado",
-  "i18n:a11y:keyboard-shortcuts:caret-extension": "Esta función requiere el uso de una <strong>extensión de Caret Browsing</strong> que le permite navegar y seleccionar el texto de páginas web con las flechas del teclado.",
+  "i18n:a11y:keyboard-shortcuts:caret-browsing": "Para seleccionar texto y tomar notas, active la navegación por cursor presionando F7 (o Fn+F7) en su navegador. Esto le permitirá navegar y seleccionar texto usando las flechas del teclado, sin necesidad de ninguna extensión.",
   "i18n:a11y:keyboard-shortcuts:move-focus-through-each-line": "Navegar el foco por cada línea de texto",
   "i18n:a11y:keyboard-shortcuts:select-text": "Seleccionar texto",
   "i18n:a11y:keyboard-shortcuts:create-highlight-or-note": "Resaltar o crear una nota después de seleccionar el texto",
