+++ src/app/messages/pl/messages.json
@@ -2,7 +2,6 @@
   "i18n:a11y:skipToContent": "Przejdź do treści",
   "i18n:a11y:accessibilityHelp": "Przejdź do informacji o dostępności",
   "i18n:a11y:accessibilityHelp:link": "https://openstax.org/accessibility-statement",
-  "i18n:a11y:keyboard-shortcuts:menu": "Menu skrótów klawiszowych",
   "i18n:a11y:keyboard-shortcuts:heading": "Skróty klawiszowe",
   "i18n:a11y:keyboard-shortcuts:close": "Kliknij, aby zamknąć menu skrótów klawiszowych",
   "i18n:a11y:keyboard-shortcuts:sub-heading": "Nawigacja z użyciem klawiszy",
@@ -15,7 +14,7 @@
   "i18n:a11y:keyboard-shortcuts:deselect-highlight-color": "Odznacz bieżący kolor podświetlenia",
   "i18n:a11y:keyboard-shortcuts:save-or-cancel": "Zachowaj albo odrzuć zmiany w notatce",
   "i18n:a11y:keyboard-shortcuts:creating-highlights-and-notes": "Tworzenie zakreśleń i notatek za pomocą skrótów klawiszowych",
-  "i18n:a11y:keyboard-shortcuts:caret-extension": "Funkcjonalność wymaga użycia <strong>rozszerzenia Caret Browsing</strong>, które umożliwia przeglądanie i zaznaczanie tekstu na stronach internetowych za pomocą klawiszy ze strzałkami.",
+  "i18n:a11y:keyboard-shortcuts:caret-browsing": "Aby zaznaczać tekst i robić notatki, włącz nawigację kursorem, naciskając F7 (lub Fn+F7) w swojej przeglądarce. Pozwala to przeglądać i zaznaczać tekst za pomocą klawiszy strzałek, bez potrzeby instalowania żadnego rozszerzenia.",
   "i18n:a11y:keyboard-shortcuts:move-focus-through-each-line": "Aktywuj fragment tekstu",
   "i18n:a11y:keyboard-shortcuts:select-text": "Zaznacz tekst",
   "i18n:a11y:keyboard-shortcuts:create-highlight-or-note": "Stwórz zakreślenie lub notatkę po zaznaczeniu tekstu",
